==> 3 -- Combine Near Miss Sequences Dataframe.ipynb <==
# Step 3: Combine Near Miss Sequences Dataframes

# Imports
import pandas as pd

## Load Dataframes
- Can only combine 2 at a time
- If need to combine more, do it sequencially i.e. combine the first 2, and then use the output to combine with the 3rd
- Set the bigger (more segments) dataframe to df1
- Set the smaller (less segments) dataframe to df2

# Load the first dataframe (Let this be the 'BASE' dataframe: We are adding to this)
df1 = pd.read_csv('data/df_near_miss_windows_fingerprint.csv')

# Load the second dataframe
df2 = pd.read_csv('data/df_near_miss_windows_manuevers.csv')


## Data Formatting
- Sort out the seg_num_id so that there aren't duplicates for different near miss sequences



# Identify the largest seg_num_id in the base dataframe
largest_seg_num_id = df1['seg_num_id'].max()

# Use this to manipulate the seg_num_id of the second dataframe
df2['seg_num_id'] = df2['seg_num_id'] + largest_seg_num_id


## Combine the two dataframes
- Concatonate the two dataframes


df_combined = pd.concat([df1, df2], ignore_index=True)


# drop vin and veh_assist mode columns if it exists in df_combined
# Drop 'vin' column if it exists
if 'vin' in df_combined.columns:
    df_combined = df_combined.drop(columns=['vin'])

# Drop 'veh_assist_mode' column if it exists
if 'veh_assist_mode' in df_combined.columns:
    df_combined = df_combined.drop(columns=['veh_assist_mode'])


## Save the combined Near Miss Sequences
- Make sure name this based on the dataframes that were combined

df_combined.to_csv('data/df_near_miss_windows_combined1.csv', index=False)
